feat: visual updates, mainly in docs
