+++ src/screens/Docs/Pages/About/scenes/FAQ/index.tsx
